fintrek-async/requirements.txt
# FastAPI и сервер
#fastapi==0.104.1
#uvicorn[standard]==0.24.0

# База данных и ORM (ASYNC)
#sqlalchemy[asyncio]==2.0.23
#asyncpg==0.29.0
#alembic==1.12.1

# Валидация и настройки
#pydantic==2.5.0
#pydantic-settings==2.1.0

# Безопасность
#python-jose[cryptography]==3.3.0
#passlib[bcrypt]==1.7.4
#cryptography==41.0.7
#python-multipart==0.0.6

# Redis и кэширование (ASYNC)
#redis[hiredis]==5.0.1

# HTTP клиент для работы с внешними API (ASYNC)
#httpx==0.25.2
#aiohttp==3.9.1

# Утилиты
#python-dotenv==1.0.0
#email-validator==2.1.0

# Тестирование
#pytest==7.4.3
#pytest-asyncio==0.21.1
#httpx==0.25.2



fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.7
python-multipart>=0.0.6
redis[hiredis]>=5.0.1
fastapi-cache2>=0.2.1
httpx>=0.25.2
aiohttp>=3.9.1
python-dotenv>=1.0.0
email-validator>=2.1.0
pytest>=7.4.3
pytest-asyncio>=0.21.1